backend/requirements.txt
fastapi==0.98.0
pydantic==1.10.2
starlette==0.28.0
uvicorn==0.22.0
redis
uuid
argon2-cffi